requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy<2.0.0
torch==2.4.1+cpu
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
newsapi-python>=0.2.7
tweepy>=4.14.0

# LLM & Embeddings
openai>=1.0.0
sentence-transformers>=2.2.2
firebase-admin>=6.2.0


# Vector DB
faiss-cpu>=1.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web & Delivery
fastapi>=0.104.0
python-multipart>=0.0.6
uvicorn>=0.24.0
python-telegram-bot>=20.0
jinja2>=3.1.2
gTTS>=2.4.0
twilio>=8.0.0
aiohttp>=3.9.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pydantic>=2.0.0
loguru>=0.7.0
pytest>=7.0.0
